requirements.txt
PySide6>=6.6.0,<7.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
imagehash>=4.3.0
mediapipe>=0.10.0
send2trash>=1.8.0
